+++ threedigrid/admin/lines/models.py
@@ -9,6 +9,8 @@ For an overview of the kcu types see :ref:`kcu-label`.
 """
 
 
+import warnings
+
 import numpy as np
 
 from threedigrid.admin.lines import exporters, subsets
@@ -89,6 +91,11 @@ class Lines(Model):
 
     @property
     def breaches(self):
+        warnings.warn(
+            "lines.breaches is going to be removed "
+            " in the near future, please use the root breaches instead",
+            UserWarning,
+        )
         return self._filter_as(Breaches, kcu=55)
 
     @property